componetization
